requirements.txt
openai
pytesseract
PyPDF2
python-docx
gpt-index